improve forum icons
