+++ src/components/Screens/Forum/Thread/ForumThreadScreen.tsx
@@ -12,8 +12,6 @@ import {MaterialHeaderButton} from '../../../Buttons/MaterialHeaderButton';
 import {AppIcons} from '../../../../libraries/Enums/Icons';
 import {useUserData} from '../../../Context/Contexts/UserDataContext';
 import {ForumThreadScreenActionsMenu} from '../../../Menus/Forum/ForumThreadScreenActionsMenu';
-import {useForumRelationMutation} from '../../../Queries/Forum/ForumRelationQueries';
-import {useAppTheme} from '../../../../styles/Theme';
 import {useTwitarr} from '../../../Context/Contexts/TwitarrContext';
 import {ForumPostListActions} from '../../../Reducers/Forum/ForumPostListReducer';
 import {ListTitleView} from '../../../Views/ListTitleView';
@@ -45,8 +43,6 @@ export const ForumThreadScreen = ({route, navigation}: Props) => {
   const [refreshing, setRefreshing] = useState(false);
   const {forumData, setForumData, forumPosts, dispatchForumPosts, forumListData, dispatchForumListData} = useTwitarr();
   const {profilePublicData} = useUserData();
-  const relationMutation = useForumRelationMutation();
-  const theme = useAppTheme();
   const postFormRef = useRef<FormikProps<PostContentData>>(null);
   const postCreateMutation = useForumPostCreateMutation();
   const {setErrorMessage} = useErrorHandler();
@@ -85,62 +81,6 @@ export const ForumThreadScreen = ({route, navigation}: Props) => {
     }
   };
 
-  const handleFavorite = useCallback(() => {
-    if (forumData) {
-      setRefreshing(true);
-      relationMutation.mutate(
-        {
-          forumID: forumData.forumID,
-          relationType: 'favorite',
-          action: forumData.isFavorite ? 'delete' : 'create',
-        },
-        {
-          onSuccess: () => {
-            setForumData({
-              ...forumData,
-              isFavorite: !forumData.isFavorite,
-            });
-            dispatchForumListData({
-              type: ForumListDataActions.updateRelations,
-              forumID: forumData.forumID,
-              isMuted: forumData.isMuted,
-              isFavorite: !forumData.isFavorite,
-            });
-          },
-          onSettled: () => setRefreshing(false),
-        },
-      );
-    }
-  }, [dispatchForumListData, forumData, relationMutation, setForumData]);
-
-  const handleMute = useCallback(() => {
-    if (forumData) {
-      setRefreshing(true);
-      relationMutation.mutate(
-        {
-          forumID: forumData.forumID,
-          relationType: 'mute',
-          action: forumData.isMuted ? 'delete' : 'create',
-        },
-        {
-          onSuccess: () => {
-            setForumData({
-              ...forumData,
-              isMuted: !forumData.isMuted,
-            });
-            dispatchForumListData({
-              type: ForumListDataActions.updateRelations,
-              forumID: forumData.forumID,
-              isMuted: !forumData.isMuted,
-              isFavorite: forumData.isFavorite,
-            });
-          },
-          onSettled: () => setRefreshing(false),
-        },
-      );
-    }
-  }, [dispatchForumListData, forumData, relationMutation, setForumData]);
-
   const getNavButtons = useCallback(() => {
     // Typescript struggles
     if (!forumData) {
@@ -152,17 +92,6 @@ export const ForumThreadScreen = ({route, navigation}: Props) => {
       <View>
         <HeaderButtons HeaderButtonComponent={MaterialHeaderButton}>
           <Item title={'Reload'} iconName={AppIcons.reload} onPress={onRefresh} />
-          {forumData && (
-            <Item
-              title={'Pinned Posts'}
-              iconName={AppIcons.pin}
-              onPress={() =>
-                navigation.push(CommonStackComponents.forumPostPinnedScreen, {
-                  forumID: forumData.forumID,
-                })
-              }
-            />
-          )}
           {eventID && (
             <Item
               title={'Event'}
@@ -170,49 +99,15 @@ export const ForumThreadScreen = ({route, navigation}: Props) => {
               onPress={() => navigation.push(CommonStackComponents.eventScreen, {eventID: eventID})}
             />
           )}
-          <Item
-            title={'Favorite'}
-            color={
-              forumData.isFavorite
-                ? theme.colors.twitarrYellow
-                : forumData.isMuted
-                ? theme.colors.onSurfaceDisabled
-                : undefined
-            }
-            iconName={AppIcons.favorite}
-            onPress={handleFavorite}
-            disabled={forumData.isMuted}
+          <ForumThreadScreenActionsMenu
+            forumData={forumData}
+            setForumData={setForumData}
+            setRefreshing={setRefreshing}
           />
-          {forumData.creator.userID !== profilePublicData?.header.userID && (
-            <Item
-              title={'Mute'}
-              color={
-                forumData.isMuted
-                  ? theme.colors.twitarrNegativeButton
-                  : forumData.isFavorite
-                  ? theme.colors.onSurfaceDisabled
-                  : undefined
-              }
-              iconName={AppIcons.mute}
-              onPress={handleMute}
-              disabled={forumData.isFavorite}
-            />
-          )}
-          <ForumThreadScreenActionsMenu forumData={forumData} />
         </HeaderButtons>
       </View>
     );
-  }, [
-    forumData,
-    handleFavorite,
-    handleMute,
-    onRefresh,
-    profilePublicData?.header.userID,
-    navigation,
-    theme.colors.onSurfaceDisabled,
-    theme.colors.twitarrNegativeButton,
-    theme.colors.twitarrYellow,
-  ]);
+  }, [setForumData, forumData, onRefresh, navigation]);
 
   useEffect(() => {
     navigation.setOptions({
